ml_experiments/receipt_classifier/requirements.txt
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
mlflow==2.12.1
pyyaml==6.0.1
joblib==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
jupyterlab==4.1.8
ipykernel==6.29.4
ipython==8.24.0
